Enable safari (#25)
There was a bug that only appeared in the combination of safari and the
whisper API. Safari's MediaStream recorder would only record mp4 files,
which record and play just fine, but once sent to the whisper API, it
would just respond with garbage.

The issue and solution is fully described
[here](https://community.openai.com/t/whisper-api-cannot-read-files-correctly/93420/54?u=jonnylangefeld).

Besides this there was another bug in Safari that would always re-render
the hero component after the demo scrolled high enough. This didn't
happen in Chrome. But it turned out that if the width of the hero
component is set explicitly to `w-fit`, it works. So that's also fixed
in this PR.

And lastly on iPhone the demo component y translation was a little low
and you couldn't really see it peaking in. So I just set this a little
higher. Ideally I'll make it dynamic one day, but the way the scroll
handler works today makes that a little more complicated, so this was a
compromise.

fix #17