✨ Feat: Merge partial responses 🎉
This feature allows for partial responses. This is realized by sending the last response alongside the openAI API request. I experimented with sending the entire history of previous transcripts, but it turned out to not be necessary. We only needed to send the last response, which would be the most refined. The user can then make a request to change just one thing and openAI will keep the remaining data from the previous response

Fix #16